make spec and doc match, extract tower auth helper method
